+++ docs/CHANGELOG.rst
@@ -15,6 +15,9 @@ Added
 
 Changed
 -------
+- **BACKWARD INCOMPATIBLE:** Refactor ``upload-fasta-nucl`` process:
+  ``species`` and ``build`` input information on FASTA file upload are
+  now mandatory, while ``source`` input has been removed.
 - Add trimming with Trimmomatic in ``workflow-wgbs-single`` and
   ``workflow-wgbs-paired`` workflows
 - Make intervals an optional input in ``bqsr`` process